Make base metrics an abc, move noop metrics to noop_metrics.py
