+++ mypy.ini
@@ -33,3 +33,6 @@ disallow_untyped_defs = True
 
 [mypy-paasta_tools.metrics.metrics_lib]
 disallow_untyped_defs = True
+
+[mypy-paasta_tools.metrics.noop_metrics]
+disallow_untyped_defs = True
