+++ tox.ini
@@ -140,6 +140,7 @@ mypy_paths =
     paasta_tools/autoscaling/autoscaling_cluster_lib.py
     paasta_tools/utils.py
     paasta_tools/metrics/metrics_lib.py
+    paasta_tools/metrics/noop_metrics.py
     tests/deployd/test_common.py
     tests/deployd/test_watchers.py
     tests/frameworks
